requirements.txt
coloredlogs==15.0.1
geopy==2.4.1
pymobiledevice3==2.46.1
PyYAML==6.0.1
